Renamed the parser and validator files specifically needed for the area calculation
